+++ src/Crowd-Fund/CrowdFund.sol
@@ -1,34 +1,35 @@
 // SPDX-License-Identifier: MIT
 pragma solidity 0.8.20;
 
-import { IERC20 } from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
-import { SafeERC20 } from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
+import {IERC20} from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
+import {SafeERC20} from "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 
-import { ICrowdFund } from "./interfaces/ICrowdFund.sol";
-import { DataTypes } from "./libs/DataTypes.sol";
+import {ICrowdFund} from "./interfaces/ICrowdFund.sol";
+import {DataTypes} from "./libs/DataTypes.sol";
 
 contract CrowdFund is ICrowdFund {
-
-    /********** STATE VARIABLES **********/
+    /**
+     * STATE VARIABLES *********
+     */
     address public immutable TOKEN;
-    uint public immutable MAX_DURATION;
+    uint256 public immutable MAX_DURATION;
 
-    uint public count;
-    mapping(uint => DataTypes.Campaign) public campaigns;
-    mapping(uint => mapping(address => uint)) public pledgedAmount;
+    uint256 public count;
+    mapping(uint256 => DataTypes.Campaign) public campaigns;
+    mapping(uint256 => mapping(address => uint256)) public pledgedAmount;
 
-    /********** CONSTRUCTOR **********/
-    constructor(address _token, uint _maxDuration) {
+    /**
+     * CONSTRUCTOR *********
+     */
+    constructor(address _token, uint256 _maxDuration) {
         TOKEN = _token;
         MAX_DURATION = _maxDuration;
     }
 
-    /********** FUNCTIONS **********/
-    function launch (
-        uint _goal,
-        uint32 _startAt,
-        uint32 _endAt
-    ) external override {
+    /**
+     * FUNCTIONS *********
+     */
+    function launch(uint256 _goal, uint32 _startAt, uint32 _endAt) external override {
         if (_startAt < block.timestamp) {
             revert StartAtIsInThePast();
         }
@@ -52,7 +53,7 @@ contract CrowdFund is ICrowdFund {
         emit CampaignLaunched(count, msg.sender, _goal, _startAt, _endAt);
     }
 
-    function cancel(uint _id) external override {
+    function cancel(uint256 _id) external override {
         DataTypes.Campaign memory campaign = campaigns[_id];
         if (msg.sender != campaign.creator) {
             revert CampaignNotCreator();
@@ -65,7 +66,7 @@ contract CrowdFund is ICrowdFund {
         emit CampaignCanceled(_id);
     }
 
-    function pledge(uint _id, uint _amount) external override {
+    function pledge(uint256 _id, uint256 _amount) external override {
         DataTypes.Campaign storage campaign = campaigns[_id];
 
         if (campaign.startAt > block.timestamp) {
@@ -82,7 +83,7 @@ contract CrowdFund is ICrowdFund {
         emit Pledged(msg.sender, _id, _amount);
     }
 
-    function unpledge(uint _id, uint _amount) external override {
+    function unpledge(uint256 _id, uint256 _amount) external override {
         DataTypes.Campaign storage campaign = campaigns[_id];
 
         if (campaign.endAt < block.timestamp) {
@@ -96,7 +97,7 @@ contract CrowdFund is ICrowdFund {
         emit Unpledged(_id, msg.sender, _amount);
     }
 
-    function claim(uint _id) external override {
+    function claim(uint256 _id) external override {
         DataTypes.Campaign storage campaign = campaigns[_id];
 
         if (msg.sender != campaign.creator) {
@@ -118,7 +119,7 @@ contract CrowdFund is ICrowdFund {
         emit Claimed(_id, msg.sender, campaign.pledged);
     }
 
-    function refund(uint _id) external override {
+    function refund(uint256 _id) external override {
         DataTypes.Campaign storage campaign = campaigns[_id];
 
         if (block.timestamp < campaign.endAt) {
@@ -128,7 +129,7 @@ contract CrowdFund is ICrowdFund {
             revert CampaignPledgedAmountIsLessThanGoal();
         }
 
-        uint balance = pledgedAmount[_id][msg.sender];
+        uint256 balance = pledgedAmount[_id][msg.sender];
         pledgedAmount[_id][msg.sender] = 0;
         SafeERC20.safeTransfer(IERC20(TOKEN), msg.sender, balance);
