+++ src/Crowd-Fund/interfaces/ICrowdFund.sol
@@ -2,15 +2,19 @@
 pragma solidity 0.8.20;
 
 interface ICrowdFund {
-    /********** EVENTS **********/
-    event CampaignLaunched(uint indexed id, address indexed creator, uint goal, uint32 startAt, uint32 endAt);
-    event CampaignCanceled(uint indexed id);
-    event Pledged(address indexed pledger, uint indexed id, uint amount);
-    event Unpledged(uint indexed id, address indexed pledger, uint amount);
-    event Claimed(uint indexed id, address indexed claimer, uint amount);
-    event Refunded(uint indexed id, address indexed refundee, uint amount);
+    /**
+     * EVENTS *********
+     */
+    event CampaignLaunched(uint256 indexed id, address indexed creator, uint256 goal, uint32 startAt, uint32 endAt);
+    event CampaignCanceled(uint256 indexed id);
+    event Pledged(address indexed pledger, uint256 indexed id, uint256 amount);
+    event Unpledged(uint256 indexed id, address indexed pledger, uint256 amount);
+    event Claimed(uint256 indexed id, address indexed claimer, uint256 amount);
+    event Refunded(uint256 indexed id, address indexed refundee, uint256 amount);
 
-    /********** ERRORS **********/
+    /**
+     * ERRORS *********
+     */
     error StartAtIsInThePast();
     error EndAtIsLessThanStartAt();
     error EndAtIsGreaterThanMaxDuration();
@@ -21,11 +25,13 @@ interface ICrowdFund {
     error CampaignAlreadyClaimed();
     error CampaignNotEnded();
 
-    /********** FUNCTIONS **********/
-    function launch(uint _goal, uint32 _startAt, uint32 _endAt) external;
-    function cancel(uint _id_) external;
-    function pledge(uint _id, uint _amount) external;
-    function unpledge(uint _id, uint _amount) external;
-    function claim(uint _id) external;
-    function refund(uint _id) external;
+    /**
+     * FUNCTIONS *********
+     */
+    function launch(uint256 _goal, uint32 _startAt, uint32 _endAt) external;
+    function cancel(uint256 _id_) external;
+    function pledge(uint256 _id, uint256 _amount) external;
+    function unpledge(uint256 _id, uint256 _amount) external;
+    function claim(uint256 _id) external;
+    function refund(uint256 _id) external;
 }
