+++ src/Crowd-Fund/libs/DataTypes.sol
@@ -4,8 +4,8 @@ pragma solidity 0.8.20;
 library DataTypes {
     struct Campaign {
         address creator;
-        uint goal;
-        uint pledged;
+        uint256 goal;
+        uint256 pledged;
         uint32 startAt;
         uint32 endAt;
         bool claimed;
